correct type

+++ white_box/cifar10/save_images_cifar10.py
@@ -44,7 +44,7 @@ if __name__ == "__main__":
     parser.add_argument('--weighted', type=bool, default=False)
     parser.add_argument('--dataset', type=str, default='test')
     parser.add_argument('--first_index', type=int, default=0)
-    parser.add_argument('--last_index', type=float, default=10000)
+    parser.add_argument('--last_index', type=int, default=10000)
     args = parser.parse_args()
 
     cifar10_save_attacks(args.weighted, args.dataset, args.first_index, args.last_index)
